chore: move project schemas to separate module
