+++ codegenerator/openapi/keystone_schemas/project.py
@@ -16,6 +16,10 @@ from typing import Any
 
 from keystone.resource import schema as ks_schema
 
+from codegenerator.common.schema import TypeSchema
+from codegenerator.common.schema import ParameterSchema
+from codegenerator.openapi.keystone_schemas import common
+
 
 PROJECT_SCHEMA: dict[str, Any] = {
     "type": "object",
@@ -90,3 +94,95 @@ PROJECT_LIST_PARAMETERS = {
         "x-openstack": {"min-ver": "3.4"},
     },
 }
+
+
+def _post_process_operation_hook(
+    openapi_spec, operation_spec, path: str | None = None
+):
+    """Hook to allow service specific generator to modify details"""
+    operationId = operation_spec.operationId
+    if operationId == "projects:get":
+        for (
+            key,
+            val,
+        ) in PROJECT_LIST_PARAMETERS.items():
+            openapi_spec.components.parameters.setdefault(
+                key, ParameterSchema(**val)
+            )
+            ref = f"#/components/parameters/{key}"
+            if ref not in [x.ref for x in operation_spec.parameters]:
+                operation_spec.parameters.append(ParameterSchema(ref=ref))
+
+
+def _get_schema_ref(
+    openapi_spec,
+    name,
+    description=None,
+    schema_def=None,
+    action_name=None,
+) -> tuple[str | None, str | None, bool]:
+    mime_type: str = "application/json"
+    ref: str
+    # Projects
+    if name == "ProjectsPostRequest":
+        openapi_spec.components.schemas.setdefault(
+            name,
+            TypeSchema(**PROJECT_CREATE_REQUEST_SCHEMA),
+        )
+        ref = f"#/components/schemas/{name}"
+    elif name == "ProjectsPostResponse":
+        openapi_spec.components.schemas.setdefault(
+            name,
+            TypeSchema(**PROJECT_CONTAINER_SCHEMA),
+        )
+        ref = f"#/components/schemas/{name}"
+    elif name == "ProjectPatchRequest":
+        openapi_spec.components.schemas.setdefault(
+            name,
+            TypeSchema(**PROJECT_UPDATE_REQUEST_SCHEMA),
+        )
+        ref = f"#/components/schemas/{name}"
+    elif name == "ProjectPatchResponse":
+        openapi_spec.components.schemas.setdefault(
+            name,
+            TypeSchema(**PROJECT_CONTAINER_SCHEMA),
+        )
+        ref = f"#/components/schemas/{name}"
+    elif name == "ProjectsGetResponse":
+        openapi_spec.components.schemas.setdefault(
+            name, TypeSchema(**PROJECTS_SCHEMA)
+        )
+        ref = f"#/components/schemas/{name}"
+    elif name == "ProjectGetResponse":
+        openapi_spec.components.schemas.setdefault(
+            name,
+            TypeSchema(**PROJECT_CONTAINER_SCHEMA),
+        )
+        ref = f"#/components/schemas/{name}"
+
+    # Project Tags
+    elif name == "ProjectsTagPutRequest":
+        openapi_spec.components.schemas.setdefault(
+            name, TypeSchema(**ks_schema.project_tag_create)
+        )
+        ref = f"#/components/schemas/{name}"
+    elif name == "ProjectsTagsPutRequest":
+        openapi_spec.components.schemas.setdefault(
+            name, TypeSchema(**ks_schema.project_tags_update)
+        )
+        ref = f"#/components/schemas/{name}"
+    elif name == "ProjectsTagsGetResponse":
+        openapi_spec.components.schemas.setdefault(
+            name, TypeSchema(**common.TAGS_SCHEMA)
+        )
+        ref = f"#/components/schemas/{name}"
+    elif name == "ProjectsTagsPutResponse":
+        openapi_spec.components.schemas.setdefault(
+            name, TypeSchema(**common.TAGS_SCHEMA)
+        )
+        ref = f"#/components/schemas/{name}"
+
+    else:
+        return (None, None, False)
+
+    return (ref, mime_type, True)
